initial status for kubernetes
