+++ pkg/status/status.go
@@ -1,0 +1,9 @@
+package status
+
+import "context"
+
+type Manager interface {
+	UpdateStatus(ctx context.Context, s *Status) error
+	//UpdateSubscription()
+	//DeleteSubscription()
+}
